requirements.txt
colorama
openai
Pillow
python-dotenv
requests
simple-term-menu
Flask
tinydb